Actually, i don't remember a lot of this changes
I did a new class called TextControl to control the whiles i use to get the input from the user, but not much apart of that. It's to reduce code complexity and extention. GN! Im sick and need to rest, love you even if i don't know who is reading this